feat: Add BottomSheetComponent for displaying content in a bottom sheet
The code changes involve adding a new component called BottomSheetComponent. This component is responsible for displaying content in a bottom sheet, allowing users to interact with the content while keeping it visible on the screen. This feature enhances the user experience and provides a more intuitive interface.

Recent user commits:
- feat: Add MapPoints component for displaying points on a map
- refactor: Remove BottomSheetComponent and update dependencies
- feat: Add new image and update npm dependencies

Recent repository commits:
- new image
- chore: Update npm dependencies and add @react-native-async-storage/async-storage
- chore: Update npm dependencies and fix import statements
- chore: Update npm dependencies to latest versions
- refactor: Update LoginScreen component to use BaseFontAwesomeIcon
- Change interface to color black
- new config in ts
- delete imagenes
- Header option oculto:
- Nuevas ventanas que aparecen en el drawer, se agregaron iconos